requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
loguru
pytest
numpy<2.2.0
python-multipart
httpx
scikit-learn